Gestion de l'envoie d'email au candidat en cas de refus ou d'acceptat…

+++ app/Http/Controllers/CandidatureController.php
@@ -2,9 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Mail\DemandeDeCandidatureAccepter;
+use App\Mail\DemandeDeCandidatureRefuser;
 use App\Models\Candidature;
 use App\Models\Formation;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Validator;
 
 class CandidatureController extends Controller
@@ -14,6 +17,10 @@ class CandidatureController extends Controller
      */
     public function index()
     {
+        if (auth()->user()->role_id==2) {
+            return response()->json(['message' => 'Vous n\'avez pas les droits pour faire cette action'], 403);
+        }
+
         $candidatures=Candidature::all();
         return response()->json([
             "message"=>"Voici la listes de toutes les candidatures",
@@ -22,6 +29,10 @@ class CandidatureController extends Controller
     }
 
     public function userCandidatures(Request $request){
+        if (auth()->user()->role_id==1) {
+            return response()->json(['message' => 'Non autorisé. Seuls les candidats peuvent voir leurs candidatures.'], 403);
+        }
+
         foreach ($request->user()->candidatures as $candidature) {
             $candidatures[]=$candidature;
         }
@@ -34,6 +45,9 @@ class CandidatureController extends Controller
     
     public function accepter()
     {
+        if (auth()->user()->role_id==2) {
+            return response()->json(['message' => 'Vous n\'avez pas les droits pour faire cette action'], 403);
+        }
         $candidatures=Candidature::where('statut','accepte')->get();
         return response()->json([
             "message"=>"Voici la listes de toutes les candidatures qui on été accepté",
@@ -52,6 +66,9 @@ class CandidatureController extends Controller
 
     public function refuser()
     {
+        if (auth()->user()->role_id==2) {
+            return response()->json(['message' => 'Vous n\'avez pas les droits pour faire cette action'], 403);
+        }
         $candidatures=Candidature::where('statut','refuse')->get();
         return response()->json([
             "message"=>"Voici la listes de toutes les candidatures qui on été refusé",
@@ -72,6 +89,10 @@ class CandidatureController extends Controller
      */
     public function store(Request $request ,Formation $formation)
     {
+        // Vérifier si l'utilisateur a le rôle "candidat"
+        if (auth()->user()->role_id==1) {
+            return response()->json(['message' => 'Non autorisé. Seuls les candidats peuvent candidater.'], 403);
+        }
 
         $candidature=Candidature::create([
             "user_id"=>$request->user()->id,
@@ -87,7 +108,10 @@ class CandidatureController extends Controller
      * Display the specified resource.
      */
     public function show(Candidature $candidature)
-    {
+    {   
+        if (auth()->user()->role_id==2) {
+            return response()->json(['message' => 'Vous n\'avez pas les droits pour faire cette action'], 403);
+        }
         return response()->json([
             "message"=>"Voici la candidature que vous chercher",
             "candidature"=>$candidature
@@ -100,9 +124,13 @@ class CandidatureController extends Controller
      */
     public function edit(Candidature $candidature)
     {
+        if (auth()->user()->role_id==2) {
+            return response()->json(['message' => 'Vous n\'avez pas les droits pour faire cette action'], 403);
+        }
         $candidature->update([
             "statut"=>"refuse"  
         ]);
+        Mail::to($candidature->user->email)->send(new DemandeDeCandidatureRefuser());
         return response()->json([
             "message"=>"Le refus de la candidature a reussi",
             "candidature"=>$candidature
@@ -114,9 +142,13 @@ class CandidatureController extends Controller
      */
     public function update(Request $request, Candidature $candidature)
     {
+        if (auth()->user()->role_id==2) {
+            return response()->json(['message' => 'Vous n\'avez pas les droits pour faire cette action'], 403);
+        }
         $candidature->update([
             "statut"=>"accepte"  
         ]);
+        Mail::to($candidature->user->email)->send(new DemandeDeCandidatureAccepter());
 
         return response()->json([
             "message"=>"La candidature a bien été accepté",
@@ -129,6 +161,9 @@ class CandidatureController extends Controller
      */
     public function destroy(Candidature $candidature)
     {
+        if (auth()->user()->role_id==2) {
+            return response()->json(['message' => 'Vous n\'avez pas les droits pour faire cette action'], 403);
+        }
         $candidature->delete();
         return response()->json([
             "message"=>"Cette candidature a été bien supprimer",
